# backend/docs/openapi.yaml
basePath: /
...
info:
  description: REST API za pretragu GitHub korisnika i repozitorijuma, praćenje (watchlist)
    repozitorijuma i pregled aktivnosti (commits, issues, itd.).
  title: GitHub Activity Dashboard API
  version: 1.0.0
paths:
  /api/activity/details/{owner}/{repo}/{sha}:
    get:
      parameters:
        - example: facebook
          in: path
          name: owner
          required: true
          type: string
        - example: react
          in: path
          name: repo
          required: true
          type: string
        - description: SHA hash commit-a
          in: path
          name: sha
          required: true
          type: string
      responses:
        '200':
          description: Detalji commit-a
          schema:
            type: object
        '404':
          description: Commit nije pronađen
      summary: Detalji o pojedinačnom commit-u
      tags:
        - Activity
  /api/activity/list:
    post:
      parameters:
        - in: body
          name: body
          required: true
          schema:
            properties:
              author_filter:
                description: Filter po autoru događaja
                example: ''
                type: string
              filter:
                description: Filter po tipu događaja
                example: All
                type: string
              owner:
                example: facebook
                type: string
              repo:
                example: react
                type: string
            required:
              - owner
              - repo
            type: object
      responses:
        '200':
          description: Lista aktivnosti
          schema:
            items:
              type: object
            type: array
        '500':
          description: Greška na serveru
        '502':
          description: Greška prilikom komunikacije sa GitHub API-jem
      summary: Lista aktivnosti (događaja) za repozitorijum, uživo sa GitHub-a
      tags:
        - Activity
...
  /api/repository/details:
    post:
      parameters:
        - in: body
          name: body
          required: true
          schema:
            properties:
              url:
                description: Puno ime repozitorijuma (owner/repo) ili URL
                example: facebook/react
                type: string
              user_id:
                description: ID ulogovanog korisnika (opciono, radi logovanja)
                example: 1
                type: integer
            required:
              - url
            type: object
      responses:
        '200':
          description: Podaci o repozitorijumu i status logovanja u bazu
          schema:
            properties:
              db_status:
                type: string
              repo_data:
                type: object
            type: object
        '400':
          description: URL nije prosleđen
        '404':
          description: Repozitorijum nije pronađen na GitHub-u
        '500':
          description: Greška na serveru
      summary: Detalji repozitorijuma (uživo sa GitHub-a), sa opcionim logovanjem
        pretrage
      tags:
        - Activity
...
  /api/users/{user_id}:
    delete:
      parameters:
        - description: ID korisnika koji se briše
          in: path
          name: user_id
          required: true
          type: integer
      responses:
        '200':
          description: Korisnik uspešno obrisan
        '403':
          description: Pokušaj brisanja glavnog admina (Anja/Una) - zabranjeno
        '404':
          description: Korisnik nije pronađen
        '500':
          description: Greška na serveru
      summary: Briše korisnika po ID-u
      tags:
        - Users
    put:
      parameters:
        - description: ID korisnika koji se menja
          in: path
          name: user_id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            properties:
              username:
                example: NovoIme
                type: string
            type: object
      responses:
        '200':
          description: Ime uspešno promenjeno
        '400':
          description: Korisničko ime je već zauzeto
        '404':
          description: Korisnik nije pronađen
        '500':
          description: Greška na serveru
      summary: Ažurira korisničko ime
      tags:
        - Users
...
schemes:
  - http
  - https
swagger: '2.0'
...